requirements.txt
torch==2.8.0
torchvision==0.23.0
matplotlib==3.10.6
pillow==11.3.0